Change observed region for THEMIS Ground products
"Earth.Magnetosphere" and "Earth.NearSurface.Ionosphere" changed
to "Earth.Surface"; we had agreed in SPASE
discussions that implied regions are not what is meant, it is the
actual observed region.  Thus we don't put Energetic Particles
measured in the solar wind as having an observed region of Sun,
although they tell us about the Sun.  As it is, it is difficult to
distinguish the THEMIS spacecraft from the ground stations; choosing
region = Earth.Magnetosphere should restrict to the spacecraft.