Adjust Cluster SPASE IDs by dropping '2'
Chris Perry says that the mission is now usually referred to
as just Cluster, occasionally the name Cluster II is used,
but this tends to be in texts that are also talking about the
original mission. Cluster-2 is best avoided due to the possible
confusion with spacecraft 2. As far as the CAA is concerned all
the products that we generate use "Cluster" as the mission ID
and "Cluster-2" is only used as the observatory ID of the second
spacecraft.